Take default values into account when computing field keys (#5384)
* make GQLIntValue and GQLFloatValue hold String values (for big numbers)

* use variables and argument defaultValues in cache keys

* update test fixtures

* do nor report deprecation warnings in schema files

+++ libraries/apollo-ast/src/commonMain/kotlin/com/apollographql/apollo3/ast/internal/Lexer.kt
@@ -562,7 +562,7 @@ internal class Lexer(val src: String) {
           end = pos,
           line = line,
           column = column(start),
-          value = asString.toDouble()
+          value = asString
       )
     } else {
       Token.Int(
@@ -570,7 +570,7 @@ internal class Lexer(val src: String) {
           end = pos,
           line = line,
           column = column(start),
-          value = asString.toInt()
+          value = asString
       )
     }
   }
